стиль, обвязал DMA2D от использования с LTDC

+++ inc/paramdepend.h
@@ -1411,8 +1411,8 @@ extern "C" {
 
 	// Цветное изображение не поддерживается на этом дисплее (заглушка).
 	#define TFTRGB(r, g, b)	(0)
-	typedef uint_fast8_t COLOR_T;	/* цвета не поддерживаются - заглушка */
-	typedef uint8_t PACKEDCOLOR_T;
+	typedef uint_fast16_t COLOR_T;	/* цвета не поддерживаются - заглушка */
+	typedef uint16_t PACKEDCOLOR_T;
 
 	#define DISPLAY_BUS_INITIALIZE() do {} while (0)
